+++ src/cppNGS/VariantList.h
@@ -255,6 +255,8 @@ public:
 
 	///Adds an annotation column and returns the index of the new column.
 	int addAnnotation(QString name, QString description, QByteArray default_value="");
+	///Adds an annotation column if it does not exist and returns the index of the new/old column.
+	int addAnnotationIfMissing(QString name, QString description, QByteArray default_value="");
 
 	///Removes an annotation column by index.
 	void removeAnnotation(int index);